requirements.txt
pytest
peewee
pyreadline
pycryptodomex